Refactor validation services to use decorators, add `ObligatoryElementValidatorDecorator`, and unify validator configurations.

+++ src/main/kotlin/pl/delukesoft/portfolioserver/application/resume/model/ValidationResultDTO.kt
@@ -2,6 +2,6 @@ package pl.delukesoft.portfolioserver.application.resume.model
 
 class ValidationResultDTO(
   val isValid: Boolean,
-  val progress: Double,
+  val progress: Int,
   val validationResults: List<ValidationDomainResultDTO>
 )
